refactor add server dialog
